# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/clientes": {
      "get": {
        "operationId": "ClientesController_list",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Nome ou telefone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponsaveisDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ClientesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateResponsavelDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsavelDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/clientes/{id}": {
      "get": {
        "operationId": "ClientesController_ficha",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsavelComAnimaisDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ClientesController_updateCliente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateResponsavelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsavelDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ClientesController_removeCliente",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/clientes/{id}/animais": {
      "post": {
        "operationId": "ClientesController_addAnimal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAnimalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/animais": {
      "get": {
        "operationId": "ClientesController_buscarAnimais",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Nome do animal ou nome/telefone do tutor",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuscaAnimalDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/animais/{id}": {
      "get": {
        "operationId": "ClientesController_animal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "ClientesController_updateAnimal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAnimalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ClientesController_removeAnimal",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/animais/{id}/foto/sign-upload": {
      "post": {
        "operationId": "ClientesController_signFoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignUploadResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/animais/{id}/foto": {
      "post": {
        "operationId": "ClientesController_confirmFoto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmFotoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnimalDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/animais/{id}/vacinas": {
      "get": {
        "operationId": "ClientesController_listVacinas",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VacinaDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "ClientesController_criarVacina",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVacinaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VacinaDto"
                }
              }
            }
          }
        },
        "tags": [
          "clientes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "VETAPP API",
    "description": "API do VETAPP — versionada (/api/v1 nas próximas iterações)",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ResponsavelPetResumoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome"
        ]
      },
      "ResponsavelListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "pets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponsavelPetResumoDto"
            }
          }
        },
        "required": [
          "id",
          "nome",
          "pets"
        ]
      },
      "ListResponsaveisDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponsavelListItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateResponsavelDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Maria Silva"
          },
          "codigo": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "maria@email.com"
          },
          "telefone": {
            "type": "string",
            "example": "+55 11 99999-0000"
          },
          "documento": {
            "type": "string"
          },
          "origem": {
            "type": "string",
            "description": "Como nos conheceu?"
          }
        },
        "required": [
          "nome"
        ]
      },
      "ResponsavelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome"
        ]
      },
      "AnimalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "responsavelId": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "especie": {
            "type": "string"
          },
          "raca": {
            "type": "string"
          },
          "pelagem": {
            "type": "string"
          },
          "sexo": {
            "type": "string"
          },
          "castrado": {
            "type": "boolean"
          },
          "nascimento": {
            "type": "string"
          },
          "microchip": {
            "type": "string"
          },
          "marcacoes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pedigree": {
            "type": "boolean"
          },
          "pedigreeNumero": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "fotoUrl": {
            "type": "string",
            "description": "URL assinada (curta) da foto, se houver"
          }
        },
        "required": [
          "id",
          "responsavelId",
          "nome",
          "castrado",
          "marcacoes",
          "pedigree",
          "status"
        ]
      },
      "ClienteVendasResumoDto": {
        "type": "object",
        "properties": {
          "totalVendidoCentavos": {
            "type": "number"
          },
          "ticketMedioCentavos": {
            "type": "number"
          },
          "vendas": {
            "type": "number"
          },
          "ultimaVendaEm": {
            "type": "string"
          }
        },
        "required": [
          "totalVendidoCentavos",
          "ticketMedioCentavos",
          "vendas"
        ]
      },
      "ResponsavelComAnimaisDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          },
          "animais": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AnimalDto"
            }
          },
          "vendas": {
            "$ref": "#/components/schemas/ClienteVendasResumoDto"
          }
        },
        "required": [
          "id",
          "nome",
          "animais",
          "vendas"
        ]
      },
      "UpdateResponsavelDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "documento": {
            "type": "string"
          },
          "origem": {
            "type": "string"
          }
        }
      },
      "OkDto": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ]
      },
      "CreateAnimalDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Rex"
          },
          "codigo": {
            "type": "string"
          },
          "especie": {
            "type": "string",
            "example": "Canina"
          },
          "raca": {
            "type": "string",
            "example": "Labrador"
          },
          "sexo": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ]
          },
          "castrado": {
            "type": "boolean"
          },
          "nascimento": {
            "type": "string",
            "example": "2020-05-10",
            "description": "AAAA-MM-DD"
          },
          "pelagem": {
            "type": "string",
            "example": "Preto e marrom"
          },
          "microchip": {
            "type": "string",
            "description": "Número do microchip"
          },
          "marcacoes": {
            "description": "Marcações/tags clínicas (ex.: renal)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pedigree": {
            "type": "boolean"
          },
          "pedigreeNumero": {
            "type": "string",
            "description": "Número do pedigree"
          }
        },
        "required": [
          "nome"
        ]
      },
      "BuscaAnimalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "especie": {
            "type": "string"
          },
          "raca": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "responsavelId": {
            "type": "string"
          },
          "responsavelNome": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "nome",
          "status",
          "responsavelId",
          "responsavelNome"
        ]
      },
      "UpdateAnimalDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "codigo": {
            "type": "string"
          },
          "especie": {
            "type": "string"
          },
          "raca": {
            "type": "string"
          },
          "pelagem": {
            "type": "string"
          },
          "sexo": {
            "type": "string",
            "enum": [
              "M",
              "F"
            ]
          },
          "castrado": {
            "type": "boolean"
          },
          "nascimento": {
            "type": "string"
          },
          "microchip": {
            "type": "string"
          },
          "marcacoes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "pedigree": {
            "type": "boolean"
          },
          "pedigreeNumero": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "vivo",
              "falecido"
            ]
          }
        }
      },
      "SignUploadDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "example": "image/jpeg",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          }
        },
        "required": [
          "contentType"
        ]
      },
      "SignUploadResponseDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string"
          }
        },
        "required": [
          "key",
          "uploadUrl"
        ]
      },
      "ConfirmFotoDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          }
        },
        "required": [
          "key"
        ]
      },
      "VacinaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "animalId": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "laboratorio": {
            "type": "string"
          },
          "lote": {
            "type": "string"
          },
          "aplicadaEm": {
            "type": "string"
          },
          "proximaEm": {
            "type": "string"
          },
          "aplicadaPorNome": {
            "type": "string"
          },
          "observacao": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "animalId",
          "nome",
          "aplicadaEm"
        ]
      },
      "CreateVacinaDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "example": "Antirrábica"
          },
          "laboratorio": {
            "type": "string"
          },
          "lote": {
            "type": "string"
          },
          "aplicadaEm": {
            "type": "string",
            "example": "2026-07-08",
            "description": "AAAA-MM-DD"
          },
          "proximaEm": {
            "type": "string",
            "example": "2027-07-08",
            "description": "Próxima dose (AAAA-MM-DD)"
          },
          "observacao": {
            "type": "string"
          }
        },
        "required": [
          "nome",
          "aplicadaEm"
        ]
      },
...
    }
  }
}